Add Progress Model, Layout and Receiver

+++ app/src/main/java/de/nico_assfalg/apps/android/clausura/helper/ExamDBHelper.java
@@ -26,7 +26,7 @@ import java.io.File;
 public class ExamDBHelper extends SQLiteOpenHelper {
 
     public static final String DATABASE_NAME = "Exams.db";
-    private static final int DATABASE_VERSION = 4;
+    private static final int DATABASE_VERSION = 5;
     public static final String EXAM_TABLE_NAME = "exam";
     public static final String EXAM_COLUMN_ID = "_id";
     public static final String EXAM_COLUMN_TITLE = "title";
@@ -34,6 +34,7 @@ public class ExamDBHelper extends SQLiteOpenHelper {
     public static final String EXAM_COLUMN_DATE = "date";
     public static final String EXAM_COLUMN_TIME = "time";
     public static final String EXAM_COLUMN_NOTES = "notes";
+    public static final String EXAM_COLUMN_PROGRESS = "progress";
 
     public ExamDBHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -47,17 +48,33 @@ public class ExamDBHelper extends SQLiteOpenHelper {
                 EXAM_COLUMN_LOCATION + " TEXT, " + //Hörsaal
                 EXAM_COLUMN_DATE + " TEXT, " + //Datum im Format yyyy-mm-dd !!!
                 EXAM_COLUMN_TIME + " TEXT, " + //Uhrzeit im Format hh-mm
-                EXAM_COLUMN_NOTES + " TEXT)" //Freie, mehrzeilige Notizen
+                EXAM_COLUMN_NOTES + " TEXT, " + //Freie, mehrzeilige Notizen
+                EXAM_COLUMN_PROGRESS + " INTEGER)" //Fortschritt
         );
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + EXAM_TABLE_NAME);
-        onCreate(db);
+        if (oldVersion > newVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + EXAM_TABLE_NAME);
+            onCreate(db);
+            return;
+        }
+        //noinspection SwitchStatementWithTooFewBranches
+        switch (oldVersion) {
+            case 4:
+                db.execSQL(
+                        "ALTER TABLE " + EXAM_TABLE_NAME + " " +
+                        "ADD " + EXAM_COLUMN_PROGRESS + " INTEGER " +
+                        "DEFAULT (0)"
+                );
+                break;
+            default:
+                throw new IllegalStateException("onUpgrade() with unknown oldVersion " + oldVersion);
+        }
     }
 
-    public boolean insertExam(String title, String date, String time, String location, String notes) {
+    public boolean insertExam(String title, String date, String time, String location, String notes, int progress) {
         SQLiteDatabase db = getWritableDatabase();
         ContentValues contentValues = new ContentValues();
         contentValues.put(EXAM_COLUMN_TITLE, title);
@@ -65,11 +82,12 @@ public class ExamDBHelper extends SQLiteOpenHelper {
         contentValues.put(EXAM_COLUMN_DATE, date);
         contentValues.put(EXAM_COLUMN_TIME, time);
         contentValues.put(EXAM_COLUMN_NOTES, notes);
+        contentValues.put(EXAM_COLUMN_PROGRESS, progress);
         db.insert(EXAM_TABLE_NAME, null, contentValues);
         return true;
     }
 
-    public boolean updateExam(Integer id, String title, String date, String time, String location, String notes) {
+    public boolean updateExam(Integer id, String title, String date, String time, String location, String notes, int progress) {
         SQLiteDatabase db = getWritableDatabase();
         ContentValues contentValues = new ContentValues();
         contentValues.put(EXAM_COLUMN_TITLE, title);
@@ -77,6 +95,7 @@ public class ExamDBHelper extends SQLiteOpenHelper {
         contentValues.put(EXAM_COLUMN_DATE, date);
         contentValues.put(EXAM_COLUMN_TIME, time);
         contentValues.put(EXAM_COLUMN_NOTES, notes);
+        contentValues.put(EXAM_COLUMN_PROGRESS, progress);
         db.update(EXAM_TABLE_NAME, contentValues, EXAM_COLUMN_ID + " = ? ", new String[] { Integer.toString(id) } );
         return true;
     }
